pyproject.toml
[tool.poetry]
name = "ugs_sprint_1"
version = "0.1.0"
description = ""
authors = ["GA10"]

[tool.poetry.dependencies]
python = "^3.10"
clickhouse-driver = "^0.2.4"
backoff = "^2.2.1"
aiokafka = "^0.8.0"
fastapi = "^0.88.0"
asyncio = "^3.4.3"
uvicorn = "^0.20.0"
PyJWT = "^2.6.0"

[tool.poetry.dev-dependencies]
flake8-bugbear = "^22.6.22"
flake8-commas = "^2.1.0"
flake8-comprehensions = "^3.10.0"
flake8-eradicate = "^1.3.0"
flake8-pie = "^0.16.0"
flake8-print = "^5.0.0"
flake8-quotes = "^3.3.1"
flake8-simplify = "^0.19.2"
flake8-variables-names = "^0.0.5"
pre-commit = "^2.19.0"
dlint = "^0.13.0"
black = "^22.6.0"
isort = "^5.10.1"
yesqa = "^1.4.0"
flake8 = "^5.0.4"
pytest = "^7.1.2"
pytest-asyncio = "^0.19.0"
aiohttp = "^3.8.1"
jupyter = "^1.0.0"


[tool.black]
line-length = 120
target-version = ["py38"]
include = '\.pyi?$'
skip-string-normalization = true
exclude = '''
(
  /(
      \.venv
    | \.git
    | \.env
  )/
)
'''

[tool.isort]
multi_line_output = 3
line_length = 120
include_trailing_comma = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
